+++ src/simulation/util/Cell.java
@@ -20,7 +20,7 @@ public interface Cell<T> {
 	
 	public T getLeft();
 
-	public float getTemp();
+	public double getTemp();
 	
 	public void setTemp(float temp);
